fix: Add unit tests for serialization module
- Implement comprehensive unit tests for the serialization utilities in the advanced_alchemy.utils.serialization module.
- Cover various data types including datetime, date, time, timedelta, Decimal, UUID, Path, PurePath, Enum, and IP addresses.
- Include tests for Msgspec, Orjson, and Standard Library serializers, ensuring correct encoding and decoding behavior.
- Validate custom type encoders and encoding hooks for extensibility.
- Ensure backward compatibility and deprecation warnings for legacy serialization functions.
- Test integration with optional dependencies like NumPy and Pydantic.
