+++ pyproject.toml
@@ -162,7 +162,7 @@ test = [
   "pytest>=7.4.4",
   "pytest-asyncio>=0.23.8",
   "pytest-cov>=5.0.0",
-  "pytest-databases[postgres,oracle,cockroachdb,mssql,bigquery,spanner,mysql,minio]",
+  "pytest-databases[postgres,oracle,cockroachdb,mssql,bigquery,spanner,mysql,minio] @ git+https://github.com/litestar-org/pytest-databases.git@feat/sql-server-clientless",
   "pytest-lazy-fixtures>=1.1.1",
   "pytest-rerunfailures",
   "pytest-mock>=3.14.0",
